remove useless code

+++ maths/src/main/java/code/maths/geo/Triangle.java
@@ -113,7 +113,7 @@ public final class Triangle implements HasEdges {
         long y_ = firstPoint.getYcoords();
         y_ += secondPoint.getYcoords();
         y_ += thirdPoint.getYcoords();
-        return new CompactPlanePoint(new PairNumber<Long,Long>(x_, y_),(long) NB_POINTS);
+        return new CompactPlanePoint(new PairNumber<Long,Long>(x_, y_),NB_POINTS);
     }
 
     public CompactPlanePoint getCircumCenter() {
